feat(avro): replace hamba/avro with twmb/avro v1.5.0 (#871)
Drop github.com/hamba/avro/v2 in favor of github.com/twmb/avro v1.5.0
for manifest reading, writing, and partition schema construction,
replacing the archived hamba/avro.

# Behavior differences visible to callers

- DataFile.Partition() returns int32 for plain integer partition fields
(was int). Correct per Avro spec — Avro int is 32-bit signed. Go code
doing v.(int) on partition values needs to switch to v.(int32).

- NewDataFileBuilder takes map[int]string instead of
map[int]avro.LogicalType. The corresponding field on dataFile and the
hasFieldToIDMap interface also now use string. Logical type constants
live in the new github.com/twmb/avro/atype subpackage (atype.Date,
atype.TimestampMicros, etc.), untyped string constants, drop-in wherever
a string is expected.

All other hamba/avro types used by the public surface are replaced:
avro.Schema → *avro.Schema on the handful of exported helpers that
referenced them; avro.SchemaCache (internal.AvroSchemaCache) →
map[string]*avro.Schema (internal.AvroSchemaMap).

# Dependency footprint

Removes five transitive dependencies: json-iterator/go,
modern-go/concurrent, modern-go/reflect2, ettle/strcase, and
go-viper/mapstructure (hamba's chain). twmb/avro's only direct
dependency is klauspost/compress, which iceberg-go already transitively
requires.

# Tests

New TestPartitionTypeToAvroSchemaDuplicateNamedTypes covers partition
spec scenarios with multiple fields of the same named Avro type (two
UUIDs, two same-size Fixed, two same-precision Decimal).

+++ table/table_test.go
@@ -666,7 +666,7 @@ func (t *TableWritingTestSuite) TestAddFilesPartitionedTable() {
 
 		for _, e := range entries {
 			t.Equal(map[int]any{
-				1000: 123, 1001: 650,
+				1000: int32(123), 1001: int32(650),
 			}, e.DataFile().Partition())
 		}
 	}
